TestByAutomation

+++ tools/TestAzAccounts/TestAzAccountsManagedIdentity.ps1
@@ -13,24 +13,37 @@ $resourceGroupName = 'AzAccountsTest'
 $automationAccount = 'AzAccountsTestAutomation'
 $location = 'eastus'
 $userManagedIdentity = 'AzAccountsTestUMI'
-$runbookName = 'AzAccountsTestRunbook'
 
 Set-AzContext -SubscriptionName $subscription
 New-AzResourceGroup -Name $ResourceGroupName -Location $Location
 
-. "$PSScriptRoot/DeployUserManagedIdentity.ps1" -Tenant $tenantId -Subscription $subscription -ResourceGroupName $resourceGroupName`
-                                                -UserManagedIdentity $userManagedIdentity -Location $location
+. "$PSScriptRoot/DeployRunbook.ps1" -AutomationAccount $automationAccount -IdentityType @('SystemAssigned', 'UserAssigned') -UserManagedIdentity $userManagedIdentity -ResourceGroupName $resourceGroupName -Subscription $subscription -Location $location
+Set-AzAutomationModule -AutomationAccountName $automationAccount -Name 'Az.Accounts' -ContentLinkUri $ModulePath -ContentLinkVersion $ModuleVersion -ResourceGroupName $resourceGroupName
 
-. "$PSScriptRoot/DeployRunbook.ps1" -AutomationAccount $automationAccount -UserManagedIdentity $userManagedIdentity`
-                                    -ResourceGroupName $resourceGroupName -Subscription $subscription -Location $location
-
+$runbookName = 'AzAccountsTestRunbook'
 $params = [ordered]@{
     "ResourceGroupName" = $resourceGroupName;
     "UserManagedIdentity" = $userManagedIdentity;
     "AutomationAccount" = $automationAccount;
     "Method" = "UA"
 }
-Set-AzAutomationModule -AutomationAccountName $automationAccount -Name 'Az.Accounts' -ContentLinkUri $ModulePath -ContentLinkVersion $ModuleVersion -ResourceGroupName $resourceGroupName
 Start-AzAutomationRunbook -AutomationAccountName $automationAccount -Name $runbookName -ResourceGroupName $resourceGroupName -Parameters $params
+exit
+$output = $null
+while ($null-eq $output)
+{
+    $job = Get-AzAutomationJob -RunbookName $runbookName -AutomationAccountName $automationAccount -ResourceGroupName $resourceGroupName
+    $output = Get-AzAutomationJobOutput -Id $job.JobId -AutomationAccountName $automationAccount -ResourceGroupName $resourceGroupName -Stream 'Any'
+    Start-Sleep -Seconds 5
+}
 
-Remove-AzAutomationAccount -Name $automationAccount -ResourceGroupName $resourceGroupName -Force
+$errOutput = $output | Where-Object {$_.Type -eq 'Error'}
+if ($null -eq $errOutput -or $errOutput.Count -eq 0)
+{
+    Write-Output "Job run successfully!"
+    Remove-AzAutomationAccount -Name $automationAccount -ResourceGroupName $resourceGroupName -Force
+}
+else
+{
+    Write-Output "Job error count: $($errOutput.Count)"
+}
